requirements.txt
altair
pandas
streamlit
moviepy
openai
pyannote.audio
pydub
torch
yt_dlp
moviepy
cmake
sentencepiece
openai-whisper @ git+https://github.com/openai/whisper.git@517a43ecd132a2089d85f4ebc044728a71d49f6e
yt-dlp
ffmpeg-python
ffmpy
demucs
f5-tts
librosa
noisereduce
scipy
numpy
langgraph
langchain-openai